requirements.txt
appium-python-client
selenium
requests
beautifulsoup4
